Remove ElasticManager field printing_kwargs
The name of the field is confusing. It's much cleaner for the users to call
`get_connect_cmd` directly. Workers may be added in different ways to the same
ElasticManager and require a different startup. Also, `get_connect_cmd` was not
using printing_kwargs for it's kwargs by default, which was inconsistent.